Fix macos crash (#1746)
Fixed #1707
Problem description: dragging an element using QDrag->exec() crashed the program, because by the time DraggableElement was called, it automatically managed to free itself.
Even though the palette has been updated with Qt::QueuedConnection, which made it possible to call mousePressEvent
using drag->exec before cleaning the elements, the drag->exec call is not blocked in macOS. As a solution, it was possible to block the event loop for drag->exec, but instead it was decided to stupidly emit to update the palette after the corresponding operations (since cleaning is put after mousePressEvent in the event loop).

P.S. No problems with memory leaks of mPaletteElements and mPaletteItems were found, they are cleaned when QTreeWidget::clear() is called. However, they have problems with hashing.