Validate polygon before creating box2D object (#1968)
Although [PR](https://github.com/trikset/trik-studio/pull/1962) is correct, it has side effects that were not found in the tests. When using a `2D model` in the `UI`, many objects have a fixed size, and some support resize. For example, when a `wall` is resized, the `2D object` is recreated. However, when the user initially drags the wall object (before it is extended in the resize), a very small wall fragment is already created (`TwoDModelScene::mousePressEvent`), which has negative consequences when calculating the polygon. It is proposed to filter out such polygons.